Add heading highlighting test cases
If the input comes from "git log -p" (with diffs for each log entry),
then we should highlight all commit lines using a blue background
extending past the end of the line.

We do not need to highlight the first (commit) line because:
* It's always on top of the file, no navigation help needed
* If we ignore that one it becomes trivial to tell whether or not the
  input contains diffs while we're filtering.

This is to help people see where each commit starts when scrolling up
and down in the log output.