Add macports moar +pager variant support
MacPorts now has a variant of the `moar` port which installs `moar` as
`moar-pager`. We are currently applying this patch during builds but
would prefer to avoid patching upstream builds if possible. This will
also prevent any other packaging system from needing to maintain
a similar patch should the take the approach that we have or something
similar.