Change default unchanged style to Yellow
I believe this is much more helpful than red-green, so this will be the
new default.

To revert back to the old behavior, set...

    RIFF=--unchanged-style=red-green

... in your environment.

With this release we also drop the "experimental" option, which was
identical to Yellow. The old experimental / yellow is the new default,
and no options are needed.